% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils4.data.table.R
\name{convert_cols.dtb}
\alias{convert_cols.dtb}
\title{Apply convertion functions to multiple columns of a \verb{data.table} in place.}
\usage{
convert_cols.dtb(dtb, ...)
}
\arguments{
\item{dtb}{- \code{\link{data.table}} instance}

\item{...}{- one or more lists of the form \code{list(c('col1','col2'), cvt_fn, optional_args...)}
where \code{cvt_fn} is the conversion function that takes a column as its 1st argument
and additional args can be supplied via \code{optional_args}}
}
\value{
modified in place \code{dtb} is returned.
}
\description{
Convenience function to convert \code{\link{data.table}} columns in place
}
