% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils4.data.table.R
\name{fread.dtb}
\alias{fread.dtb}
\title{\verb{data.table::fread} variant that reads compressed files.}
\usage{
fread.dtb(file, ...)
}
\arguments{
\item{file}{path to a CVS file. If it has extensions \code{.gz} or \code{.bz2} (case-insensitive)
decompress on a fly.}

\item{...}{other \code{\link[data.table]{fread}} parameters.}
}
\value{
\code{data.table} instance.
}
\description{
Convenience wrapper around \code{\link[data.table]{fread}} that automatically deals with compressed
files. Currently supported are gzip and bzip2 compressions.
}
\examples{
\dontrun{
dtb <- fread.dtb('foo.bz2', colClasses='character')
}

}
