% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.naf}
\alias{is.naf}
\title{Return (scalar or vector) TRUE if elements of its arg are NA or FALSE}
\usage{
is.naf(x)
}
\arguments{
\item{x}{R object (scalar or vector)}
}
\value{
R object (scalar or vector)
}
\description{
Return (scalar or vector) TRUE if elements of its arg are NA or FALSE
}
