% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LR.find_file}
\alias{LR.find_file}
\title{Find files by regex pattern.}
\usage{
LR.find_file(dir = ".", pattern = "", index = "last",
  ignore.case = TRUE, recursive = FALSE)
}
\arguments{
\item{dir}{directory or array of directories to search in}

\item{pattern}{regex matching file name}

\item{index}{what to return. \code{"last"} for last found file, \code{"first"} for the first,
\code{"all"} for all found files; numeric index or array of indexes
to subset vector of matching files.}

\item{ignore.case}{boolean, whether to ignore the case in regex matching}

\item{recursive}{boolean, whether perform recursive search over sub-directories.}
}
\value{
NULL is no match found, otherwise path to found files as prescribed by \code{index} settings.
}
\description{
Convenience function finding files by regex pattern.
}
