% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{LR.install.packages}
\alias{LR.install.packages}
\title{Wrapper around \code{\link{install.packages}} to install tests by default.}
\usage{
LR.install.packages(pkgs, INSTALL_opts = NULL, ...)
}
\arguments{
\item{pkgs}{character vector of the names of packages whose current versions should be downloaded
from the repositories.}

\item{INSTALL_opts}{Optional character vector of additional options. \code{"--install-tests"} will be
automatically prepended. For the rest, see \code{\link{install.packages}}.}

\item{...}{The rest of parameters are passed directly to \code{\link{install.packages}}.}
}
\value{
The results of calling \code{\link{install.packages}}.
}
\description{
An option \code{"--install-tests"} is automatically prepended to \code{INSTALL_opts}
}
