% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{LR.list_package_files}
\alias{LR.list_package_files}
\title{LR.list_package_files  List all file in an installed package.}
\usage{
LR.list_package_files(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{package to consider}

\item{lib.loc}{a character vector describing the location of R library trees to search through,
or NULL. The default value of NULL corresponds to checking the loaded namespace,
then all libraries currently known in .libPaths().}
}
\value{
List of files
}
\description{
LR.list_package_files  List all file in an installed package.
}
\examples{

}
