% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj.R
\name{LR.prj_root}
\alias{LR.prj_root}
\title{Find the project root}
\usage{
LR.prj_root(dir = NULL, marker = ".PRJ_ROOT")
}
\arguments{
\item{dir}{where to start the search if a string. If \code{NULL} or \code{FALSE} then start at the
current working directory (getwd()). If \code{TRUE} (should only be used in scripts that are
sourced/executed) then try directory of the file being sourced, otherwise current working directory.}

\item{marker}{name of the project root marker file.}
}
\value{
project root or \code{NULL} if failed
}
\description{
prj_root marker file (.PRJ_ROOT by default) must exist in the project root directory.
}
