% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj.R
\name{LR.prj_skeleton}
\alias{LR.prj_skeleton}
\title{Create standard project skeleton sub-directores}
\usage{
LR.prj_skeleton(root = ".", marker = ".PRJ_ROOT", mode = "0750")
}
\arguments{
\item{root}{project root}

\item{marker}{project root marker}

\item{mode}{mode for the directories}
}
\value{
nothing
}
\description{
For project directory structure see \code{\link{LR.prj_path}}.
}
