% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_root.R
\name{LR_get.prj_root}
\alias{LR_get.prj_root}
\title{Find the project root}
\usage{
LR_get.prj_root(dir = getwd(), marker = ".PRJ_ROOT")
}
\arguments{
\item{dir}{where to start the search.}

\item{marker}{name of the project root marker file.}
}
\value{
project root or NULL if failed
}
\description{
Find the project root
}
