% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils4.data.table.R
\name{setnames.dtb}
\alias{setnames.dtb}
\title{Conveniently set column names of a \verb{data.table}.}
\usage{
setnames.dtb(dtb, rename.map)
}
\arguments{
\item{dtb}{-- \code{data.table} object}

\item{rename.map}{-- named array of the form \code{c('new1=old1',...)}}
}
\value{
The input is modified by reference, and returned (invisibly)
so it can be used in compound statements.
}
\description{
Convenience wrapper around \code{\link[data.table]{setnames}} function.
}
