% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_colnames}
\alias{std_colnames}
\title{Standardize column names to contain only low-case alphanumeric chars.}
\usage{
std_colnames(names)
}
\arguments{
\item{names}{A vector of column names}
}
\value{
A vector of standardized column names.
}
\description{
Standardize vector of column names by converting latters to low case and replacing anything
that matches regex \verb{'[^a-z0-9_\%#$@.]+'} with \verb{'_'}.
}
